New -verbose options, ErrorAction, more comments
I added -ErrorAction SilentlyContinue to each of the pieces of code that remove AppXPackages/AppXProvisionedPackages so that the screen doesn't fill up with errors. I then added a -verbose option to each of those so that you can see specifically what the script is trying to remove.

I added -verbose to most of the code that makes changes, such as creating folders, changing registry keys, and removing the bloatware.

I also added more comments to explain further what each piece of code does. On top of that I added Write-Output "" with information so that when the script is creating the folders for backing up specific registry keys it won't seem as if it is empty information.